+++ packages/uniforms/src/SimpleSchemaBridge.js
@@ -97,22 +97,18 @@ export default class SimpleSchemaBridge extends Bridge {
             error.details &&
             error.details.find &&
             error.details.find(error => error.name === name)
-        );
+        ) || null;
     }
 
     getErrorMessage (name, error) {
         const scopedError = this.getError(name, error);
-        if (scopedError) {
-            return this.schema.messageForError(
-                scopedError.type,
-                scopedError.name,
-                null,
-                scopedError.details &&
-                scopedError.details.value
-            );
-        }
-
-        return '';
+        return !scopedError ? '' : this.schema.messageForError(
+            scopedError.type,
+            scopedError.name,
+            null,
+            scopedError.details &&
+            scopedError.details.value
+        );
     }
 
     getErrorMessages (error) {